compiler error fixed

+++ src/drivers/gnss/septentrio/septentrio.cpp
@@ -1275,12 +1275,12 @@ int SeptentrioDriver::process_message()
 				for (int i = 0; i < math::min(rf_status.n, static_cast<uint8_t>(sizeof(rf_status.rf_band) / sizeof(rf_status.rf_band[0]))); i++) {
 					InfoMode status = rf_status.rf_band[i].info_mode;
 
-					if(status == InfoMode::Interference){
+					if(status == static_cast<uint8_t>(InfoMode::Interference)){
 						_message_gps_state.jamming_state = sensor_gps_s::JAMMING_STATE_DETECTED;
 						break; // Worst case, we don't need to check the other bands
 					}
 
-					if(status == InfoMode::Suppressed || status == InfoMode::Mitigated){
+					if(status == static_cast<uint8_t>(InfoMode::Suppressed) || status == static_cast<uint8_t>(InfoMode::Mitigated)){
 						_message_gps_state.jamming_state = sensor_gps_s::JAMMING_STATE_MITIGATED;
 					}
 				}
